# src/Output/openapi.yaml
openapi: 3.0.0
info:
  title: 'Calendar App'
  description: 'This API is designed to facilitate the management of a comprehensive calendar application, encompassing a broad range of functionalities including user management, event scheduling, course allocation, group coordination, department organization, and timetable structuring. It aims to provide an integrated platform for educational institutions to efficiently organize and access academic and administrative information.'
  version: 1.0.3
servers:
  -
    url: 'http://localhost/web_project/src/'
paths:
...
  '/users/{id}':
    get:
      tags:
        - Users
      summary: 'Get a user by ID'
      description: 'Retrieves a user by its ID.'
      operationId: c551a1ac1be37485844fea99fd5c4ec5
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
    put:
      tags:
        - Users
      summary: 'Update a user by ID'
      description: 'Updates a user by its ID. Only certain fields can be updated.'
      operationId: 4e73057119289c1852c31b99ead70192
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'User object that needs to be updated'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'Bad request' }
                type: object
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
    delete:
      tags:
        - Users
      summary: 'Delete a user by ID'
      description: 'Deletes a user by its ID.'
      operationId: 2e3cbd202fb5b06e4e385ea633ba7f06
      parameters:
        -
          name: id
          in: path
          description: 'ID of the user to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'User deleted successfully'
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'User not found' }
                type: object
components:
  schemas:
...
    User:
      title: User
      description: 'User entity schema'
      required:
        - id
        - email
        - firstName
        - lastName
        - type
      properties:
        id:
          description: 'Unique identifier for the user'
          type: integer
          format: int64
        email:
          description: 'User email address'
          type: string
          format: email
        firstName:
          description: "User's first name"
          type: string
        lastName:
          description: "User's last name"
          type: string
        type:
          description: 'The type of user'
          type: string
          enum:
            - student
            - teacher
            - admin
        password:
          description: 'User password, not generally returned in responses for security reasons'
          type: string
          writeOnly: true
          example: examplePassword123!
      type: object